requirements.txt
fastapi==0.100.0
onnxruntime==1.15.1
Pillow==10.0.0
pydantic==2.0.3
Requests==2.31.0
torch==2.0.1
torchvision==0.15.2